Add missing fields and some doc to PBXFileSystemSynchronizedBuildFileExceptionSet
