+++ lib/xcodeproj/project/object.rb
@@ -534,4 +534,4 @@ require 'xcodeproj/project/object/root_object'
 require 'xcodeproj/project/object/target_dependency'
 require 'xcodeproj/project/object/reference_proxy'
 require 'xcodeproj/project/object/file_system_synchronized_root_group'
-require 'xcodeproj/project/object/file_system_synchronized_build_file_exception_set'
+require 'xcodeproj/project/object/file_system_synchronized_exception_set'
